Add `Shape` trait and two implementations
Now the `core` crate has a purpose again. The `Shape`
trait is a nice abstraction, but it can be dangerous:
if we by accident, switch to dynamic dispatch, the
performance will die. We really have to be clever about
it in order to tell Rust to monomorphize the important
functions, even if we are using some kind of dynamic
dispatch further up. For now, it's simple static
dispatch.