Switch back to scalar rotate impl
The SIMD one is slower than the scalar one. It might be that this code
just maps very bad to SIMD instructions and due to amount of shuffling
and horizontal operations it becomes very slow. It might be possible
to optimize still but for now I've given up on that. More, it's
probably just more useful to actually calculate 4 or 8 DE at once using
SIMD.